+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ----------------
 
 - Add filter response curves and filters module.
+- Add pyyaml required dependency (for reading filter curves).
 
 0.2 (2015-11-15)
 ----------------